requirements.txt
grpcio
grpcio-tools
asyncio
googleapis-common-protos